requirements.txt
# Основные зависимости
fastapi>=0.102.0,<0.103.0
uvicorn>=0.23.0,<0.24.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
email-validator>=2.0.0,<3.0.0

# Работа с базой данных
sqlalchemy>=2.0.0,<3.0.0
psycopg>=3.1.10,<4.0.0
asyncpg>=0.28.0,<0.29.0
alembic>=1.12.0,<2.0.0

# Безопасность
passlib>=1.7.4,<2.0.0
bcrypt>=4.0.1,<5.0.0
python-jose>=3.3.0,<4.0.0
python-multipart>=0.0.6,<0.1.0
PyJWT>=2.5.0,<3.0.0

# Работа с файлами
boto3>=1.28.50,<2.0.0
aiofiles>=23.0.0,<24.0.0

# HTTP клиент
aiohttp>=3.8.5,<4.0.0
httpx>=0.24.1,<0.25.0

# Шаблонизатор
jinja2>=3.1.0,<3.2.0 